requirements.txt
fastapi
uvicorn
b2sdk
python-dotenv
cryptography
python-multipart